de_demo/migrations/clickhouse/0001_create_events.py
from infi.clickhouse_orm import migrations

from de_demo.apps.events.warehouse.models import (
    Events, EventsBuffer
)

operations = [
    migrations.CreateTable(Events),
    migrations.CreateTable(EventsBuffer),
]
